<feat: all Organization feature; add AvatarController>
