+++ api/src/main/java/com/aguri/captionlive/model/Organization.java
@@ -22,7 +22,7 @@ public class Organization {
 
     private String description;
 
-    private String avatar;
+    private Long avatar;
 
     @JsonIgnore
     @UpdateTimestamp
