+++ api/src/main/java/com/aguri/captionlive/service/impl/OrganizationServiceImpl.java
@@ -1,12 +1,17 @@
 package com.aguri.captionlive.service.impl;
 
 import com.aguri.captionlive.common.exception.EntityNotFoundException;
+import com.aguri.captionlive.model.FileRecord;
 import com.aguri.captionlive.model.Organization;
 import com.aguri.captionlive.repository.OrganizationRepository;
+import com.aguri.captionlive.service.FileRecordService;
 import com.aguri.captionlive.service.OrganizationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.List;
 
 @Service
@@ -44,5 +49,23 @@ public class OrganizationServiceImpl implements OrganizationService {
     public void deleteOrganization(Long id) {
         Organization organization = getOrganizationById(id);
         organizationRepository.delete(organization);
+    }
+
+    @Autowired
+    FileRecordService fileRecordService;
+
+    @Override
+    public Organization saveAvatarToStorage(Long id, MultipartFile file) {
+        System.out.println(file);
+        Organization organization = getOrganizationById(id);
+        if (!file.isEmpty()) {
+            try {
+                FileRecord fileRecord = fileRecordService.saveFile(file, "avatar" + File.separator + "organization" + File.separator + id.toString());
+                organization.setAvatar(fileRecord.getFileRecordId());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        return updateOrganization(id, organization);
     }
 }
